Working pointers, adding game.c/.h, and working on deallocating memory
